requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.22
psycopg-binary==3.1.12
psycopg==3.1.12
